standardize creation and updating of user
