+++ apps/api/v2/src/modules/users/inputs/validators/is-week-start.ts
@@ -1,0 +1,23 @@
+import { ValidatorConstraint, ValidatorConstraintInterface } from "class-validator";
+
+@ValidatorConstraint({ name: "isWeekStart", async: false })
+export class IsWeekStart implements ValidatorConstraintInterface {
+  validate(weekStart: string) {
+    if (!weekStart) return true;
+
+    const lowerCaseWeekStart = weekStart.toLowerCase();
+    return (
+      lowerCaseWeekStart === "monday" ||
+      lowerCaseWeekStart === "tuesday" ||
+      lowerCaseWeekStart === "wednesday" ||
+      lowerCaseWeekStart === "thursday" ||
+      lowerCaseWeekStart === "friday" ||
+      lowerCaseWeekStart === "saturday" ||
+      lowerCaseWeekStart === "sunday"
+    );
+  }
+
+  defaultMessage() {
+    return "weekStart must be a string either Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, or Sunday";
+  }
+}
